UCM2: Intel: sof-hda-dsp: Enable Dmic0 DRC and TDFB pre-processing
This patch adds enable of the microphone capture pre-processing
features as BootSequence initial values if the controls for them
are available. If the user later switches these off, they will
remain off.

The DRC is dynamic range control that boosts the capture level,
and the TDFB (time-domain fixed beamformer) improves capture
signal-to-noise ratio for set beam angle direction. The used
0 degrees angle is defined as angle towards notebook computer
user.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>